refactor : extension.ts 내 새로 추가한 함수 모듈화
